Add naive approach solution for part 2 with tests

+++ Day06 - Wait For It/Puzzle/BoatRecord.cs
@@ -3,8 +3,8 @@
 namespace AdventOfCode.Year2023.Day06.Puzzle;
 
 [method: SetsRequiredMembers]
-internal readonly record struct BoatRecord(int Time, int Distance)
+internal readonly record struct BoatRecord(long Time, long Distance)
 {
-	public required int Time { get; init; } = Time;
-	public required int Distance { get; init; } = Distance;
+	public required long Time { get; init; } = Time;
+	public required long Distance { get; init; } = Distance;
 }
